CVE-2008-2232: privilege escalation - ID: 2061605
afuse is vulnerable to a local privilege escalation because the expand_template
function passes metacharacters from the pathname unescaped to the shell. This
issue is tracked as CVE-2008-2232:
http://nvd.nist.gov/nvd.cfm?cvename=CVE-2008-2232

I submitted a patch to the Debian security team, and they released fixed
packages: http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=490921 However, the
vulnerability is still present in your current release and Subversion trunk.

The patch is available from the above Debian bug report, and I have also
attached it here.

Note from committer: This is from
http://sourceforge.net/tracker/?func=detail&aid=2061605&group_id=159596&atid=812464